requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
jieba-fast>=0.53.0
requests>=2.31.0
python-dotenv>=1.0.0
werkzeug==2.3.7

# Text to Speech
edge-tts>=7.0.2

# Utilities
numpy>=1.24.0
pandas==2.0.3
tqdm==4.65.0